% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_fieldAUC.R
\name{fieldAUC}
\alias{fieldAUC}
\title{Area Under the Curve}
\usage{
fieldAUC(
  data,
  trait,
  keep.columns = c("NAME", "ROW", "RANGE"),
  method = "trapezoid",
  x.start = 0,
  y.start = 0,
  frame = "long"
)
}
\arguments{
\item{trait}{one or more traits to be evaluated.}

\item{keep.columns}{columns names to be maintained in the output dataset.}

\item{method}{the type of interpolation. Can be "trapezoid" (default), "step", "linear" or "spline". More information on ??DescTools::AUC.}

\item{x.start}{value of x to start the AUC (default is 0 days after planting).}

\item{y.start}{value of y to start the AUC (default is 0).}

\item{frame}{format of output data. "long" is used for AUC values on the 1st column and traits ID on the2nd column. While "wide" is for objects with dimension n \times m where traits must be in columns and plots/sample in rows.}

\item{Data}{data table from \code{\link{fieldInfo}}.}
}
\value{
A list with a data frame with values by plot and experimental field image with format stack.
}
\description{
Calculate the area under the curve given by vectors of xy-coordinates.
}
\details{
fieldAUC
}
