% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_fieldCount.R
\name{fieldCount}
\alias{fieldCount}
\title{Calculating number of objects per plot}
\usage{
fieldCount(
  mosaic,
  fieldShape,
  value = 0,
  minSize = 0.01,
  n.core = NULL,
  pch = 16,
  cex = 0.7,
  col = "red",
  na.rm = FALSE
)
}
\arguments{
\item{mosaic}{object mask of class stack from the function \code{\link{fieldMask}}.}

\item{fieldShape}{plot shape file.}

\item{value}{referent value to vegetation pixels in the mask. If "HUE" was used on 
\code{\link{fieldMask}} the value=0.}

\item{minSize}{used to set the minimum size percentage of plant canopy  (to remove weeds and more).}

\item{n.core}{number of cores to use for multicore processing (Parallel).}

\item{pch}{point symbol, please check \code{help("points")}.}

\item{cex}{character (or symbol) expansion: a numerical vector, please check \code{help("points")}.}

\item{col}{color code or name, please check \code{help("points")}.}

\item{na.rm}{logical. Should missing values (including NaN) be removed?.}
}
\value{
A list with five element
\itemize{
  \item \code{fieldCount} is the number of objects per plot represented in DataFrame.
  \item \code{fieldShape} is the new shapeFile with stand count.
  \item \code{mosaic} is the Watershed layer.
  \item \code{objectSel} is the objects area per plot.
  \item \code{objectReject} is the objects position in the image.
}
}
\description{
The mask from function \code{\link{fieldMask}} is used to identify the number of objects per plot..
}
\details{
fieldCount
}
