% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_fieldCrop.R
\name{fieldCrop}
\alias{fieldCrop}
\title{Selecting experimental field from original image}
\usage{
fieldCrop(
  mosaic,
  fieldShape = NULL,
  nPoint = 4,
  plot = TRUE,
  remove = FALSE,
  type = "l",
  lty = 2,
  lwd = 3,
  fast.plot = FALSE
)
}
\arguments{
\item{mosaic}{object mask of class stack from the function \code{\link{fieldMask}}.}

\item{fieldShape}{crop the image using the fieldShape as reference. If fieldShape=NULL, four points should be selected 
directly on the original image to determine the experimental field.}

\item{nPoint}{number of points necessary to select field boundaries or area to remove (4 >= nPoint <= 50).}

\item{plot}{if \code{TRUE} (by default) plots the original and cropped image.}

\item{remove}{if TRUE the selected area will be removed from the image.}

\item{type}{character indicating the type of plotting, please check help("lines").}

\item{lty}{line types, please check help("lines").}

\item{lwd}{line width, please check help("lines").}

\item{fast.plot}{if TRUE only the grey scale image will be plotted as reference (faster approach).
if TRUE only the grey scale image will be plotted as reference (faster approach).}
}
\value{
A image format stack.
}
\description{
It calculates the percentage of object area in the entire mosaic or per plot using the fieldShape file.
}
\details{
fieldCrop
}
