% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_fieldDraw.R
\name{fieldDraw}
\alias{fieldDraw}
\title{Making draws (lines or polygons)}
\usage{
fieldDraw(
  mosaic,
  line = TRUE,
  ndraw = 1,
  dist = FALSE,
  distSel = 0.5,
  round = 5,
  value = 1,
  pch = 16,
  cex = 0.7,
  col = "red",
  lwd = 1
)
}
\arguments{
\item{mosaic}{object mask of class stack from the function \code{\link{fieldMask}}.}

\item{line}{if it is TRUE the selected area will be a line otherwise it will be a polygon (line=FALSE).}

\item{ndraw}{number of drawings (lines or polygons).}

\item{dist}{if it is TRUE the distance among objects or objects length will be calculated 
(mosaic should be the mask from function \code{\link{fieldMask}} and line should be line=TRUE).}

\item{distSel}{integer indicating which proportion of estimated distances should be selected (default 0.5). 
If \code{distSel = 1} all estimated distances will be saved. Some distances can be artifacts caused by the mask 
estimation and can inflate the number of distances.}

\item{round}{integer indicating the number of decimal places (round) or significant digits (signif) to be used.}

\item{value}{integer indicating the value in the mask. If value=1 the objects length will be estimated.
If \code{value = 0} the distance between objects will be estimated.}

\item{pch}{point symbol, please check help("points").}

\item{cex}{character (or symbol) expansion: a numerical vector, please check help("points").}

\item{col}{color code or name, please check help("points").}

\item{lwd}{line width, please check help("lines").}
}
\value{
A list four elements
\itemize{
  \item \code{drawData} is a matrix with each cell position "x" and "y" (line/polygon) and the cell value "layer".
  \item \code{drawObject} is the the new \code{SpatialPolygons} or \code{SpatialLines} object.
  \item \code{drawSegments} is a simplified drawData matrix with segments defined in paramter "value".
  \item \code{drawDist} is the distance or length between objects.
}
}
\description{
Extracting information (x, y, value) for all cells in a line or polygon. Also, this function can 
 be used to evaluate distances between objects or objects length.
}
\details{
fieldDraw
}
