% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_fieldHeight.R
\name{fieldHeight}
\alias{fieldHeight}
\title{Mosaic height and volume based on the digital surface model (DSM)}
\usage{
fieldHeight(dsm_before, dsm_after)
}
\arguments{
\item{dsm_before}{digital surface model (DSM) mosaic class SpatRaster to be used as base line (minumun height or soil level).}

\item{dsm_after}{digital surface model (DSM) mosaic class SpatRaster to be used as aimed line (maximum height or plants in the field).}
}
\value{
list with \code{height} and {volume} as a new \code{SpatRaster} object.
}
\description{
It calculates pixel height and volume in the entire mosaic using the digital surface model (DSM).
}
\details{
fieldHeight
}
