% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_fieldIndex.R
\name{fieldIndex}
\alias{fieldIndex}
\title{Building vegetation indices using Red, Green, Blue, Red Edge, and NIR band}
\usage{
fieldIndex(
  mosaic,
  Red = 1,
  Green = 2,
  Blue = 3,
  RedEdge = NULL,
  NIR = NULL,
  index = "HUE",
  myIndex = NULL,
  plot = TRUE
)
}
\arguments{
\item{mosaic}{object of class stack with at least 3 bands.}

\item{Red}{vector with the vegetation indices to be calculated.}

\item{Green}{user can calculate a diferent index using the bands names, e.g. "(Green+Blue)/Red-NIR/RedEdge"}

\item{Blue}{if is TRUE the original and crop image will be plotted.}

\item{RedEdge}{vector with the vegetation indices to be calculated.}

\item{NIR}{if is TRUE the original and crop image will be plotted.}

\item{index}{vector with the vegetation indices to be calculated.}

\item{myIndex}{user can calculate a diferent index using the bands names, e.g. "(Green+Blue)/Red-NIR/RedEdge"}

\item{plot}{if is TRUE the original and crop image will be plotted.}
}
\value{
A Image format stack.
}
\description{
Different vegetation indices can be calculated using at least 3 bands. 
For the list of indices please visit the FIELDimageR manual at link
}
\details{
fieldIndex
}
