% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_fieldPolygon.R
\name{fieldPolygon}
\alias{fieldPolygon}
\title{Building \code{shapefile} with polygons}
\usage{
fieldPolygon(
  mosaic,
  nPolygon = 1,
  nPoint = 4,
  polygonID = NULL,
  polygonData = NULL,
  ID = NULL,
  cropPolygon = FALSE,
  remove = FALSE,
  plot = TRUE,
  fast.plot = FALSE,
  extent = FALSE
)
}
\arguments{
\item{mosaic}{object of class stack.}

\item{nPolygon}{number of polygons.}

\item{nPoint}{number of points necessary to select field boundaries or area to remove (4 >= nPoint <= 50).}

\item{polygonID}{a vector with polygon names with same order of drawing. If is NULL the ID will be the sequence of drawing.}

\item{polygonData}{data frame with polygon ID and all attributes of each polygon (Traits as columns and polygon as rows).}

\item{ID}{the column in polygonData with polygons names (ID) which the data will be combined with fieldShape.}

\item{cropPolygon}{if TRUE the mosaic will be crooped using polygons shape.}

\item{remove}{if TRUE the selected area will be removed from the image.}

\item{plot}{if is TRUE the crop image and fieldShape will be plotted.}

\item{fast.plot}{if TRUE only the grey scale image will be plotted as reference (faster approach).}

\item{extent}{if is TRUE the entire image area will be the fieldShape (one unique plot).}
}
\value{
A list with two element
\itemize{
  \item The function returns the \code{fieldShape} format \code{SpatialPolygonsDataFrame} with plots 
  numbered by the sequence of drawings and a new reduced image with format \code{stack}. 
  The \code{polygonID} parameter can be used to identify each polygon.
}
}
\description{
The user should select points to make polygons in the image. Shapefile with polygons will be automatically built. 
 Attention: \code{fieldRotate()} is not necessary.
}
\details{
fieldPolygon
}
