% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_imgEBI.R
\name{imgEBI}
\alias{imgEBI}
\title{Converts a terra object to an EBImage object}
\usage{
imgEBI(img)
}
\arguments{
\item{img}{image/object class \code{terra}}
}
\value{
\code{EBImage} object
}
\description{
Converts a terra object to an EBImage object
}
\details{
imgEBI
}
