% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_imgLAB.R
\name{imgLAB}
\alias{imgLAB}
\title{Set min and max values and crs to terra object}
\usage{
imgLAB(img)
}
\arguments{
\item{img}{image/object class \code{terra}}
}
\value{
\code{terra} object
}
\description{
Set minmax values and crs to terra object
}
\details{
imgLAB
}
