\name{Indices.txt}
\docType{data}
\title{Vegetation indices}
\description{
  List of vegetation indices avaliable at FIELDimageR.
}
\format{Data frame with NAME, FORMULA and BANDS}
\references{
  Matias et al. (2019) FIELDimageR: A R Package to Analyze Orthomosaic Images from Agricultural Field Trials. (submitted)
}
\keyword{datasets}
