\name{RGB.rescale.R}
\docType{data}
\title{RGB rescale colors}
\description{
  Rescale continuous vector to have specified minimum and maximum.
}
\format{RGB.rescale(mosaic, num.band)}
\references{
 Matias FI, Caraza-Harter MV, Endelman JB. FIELDimageR: An R package to analyze orthomosaic images from agricultural field trials. The Plant Phenome J. 2020;e20005. https://doi.org/10.1002/ppj2.20005
}
\keyword{datasets}
