% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_fieldCount.R
\name{fieldCount}
\alias{fieldCount}
\title{Calculating number of objects per plot}
\usage{
fieldCount(
  mosaic,
  fieldShape = NULL,
  watershed = 0.05,
  plot = FALSE,
  pch = "+",
  col = "red"
)
}
\arguments{
\item{mosaic}{object mask of class stack from the function \code{\link{fieldMask}}.}

\item{fieldShape}{plot shape file.}

\item{watershed}{numeric (defaut). It is the same as \code{tolerance} parameter at \code{\link{EBImage::watershed}}. It is the minimum height of the object in the units of image intensity between its highest point (seed) and the point where it contacts another object (checked for every contact pixel). If \code{NULL} all touching objects will be considered as one polygon.}

\item{plot}{if it is TRUE the original and segmented image will be plotted with identified objects.}

\item{pch}{point symbol, please check \code{help("points")}.}

\item{col}{color code or name, please check \code{help("points")}.}
}
\value{
A list with two elements when fieldShape is provided:
\itemize{
  \item \code{fieldShape$plot_level} is the new shapeFile with objects in the plot area, perimeter, count, and mean_width.
  \item \code{fieldShape$object_level} is the new shapeFile of single objects area, perimeter, width, x and y position.
}
}
\description{
The mask from function \code{\link{fieldMask}} is used to identify the number of objects per plot..
}
\details{
fieldCount
}
