% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_fieldMap.R
\name{fieldMap}
\alias{fieldMap}
\title{Organazing the field map in function of plot order.}
\usage{
fieldMap(fieldPlot, fieldColumn, fieldRow, decreasing = FALSE)
}
\arguments{
\item{fieldPlot}{vector with plot ID}

\item{fieldColumn}{vector with column ID}

\item{fieldRow}{vector with row ID}

\item{decreasing}{if FALSE the plots will be order left to right and right to left every other line following the breeding field design.}
}
\value{
\itemize{
  \item The function returns a matrix with plots ID identified by rows and column
}
}
\description{
Field map ID identification to align with plots shapefile built with fieldShape(). The plots are numbered from left to right and top to bottom.
}
\details{
fieldMap
}
