% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_fieldObject.R
\name{fieldObject}
\alias{fieldObject}
\title{Evaluate object area, "x" distance, "y" distance, number, extent, etc.}
\usage{
fieldObject(
  mosaic,
  fieldShape = NULL,
  minArea = 0.01,
  areaValue = 0,
  watershed = FALSE,
  dissolve = TRUE,
  n.rem = 1,
  na.rm = FALSE,
  plot = TRUE
)
}
\arguments{
\item{mosaic}{object mask of class stack from the function \code{\link{fieldMask}}.}

\item{fieldShape}{evaluate the object per plot using the fieldShape as reference. 
If fieldShape=NULL, the object will be evaluated directly for the entire original image.}

\item{minArea}{used to set the minimum size percentage of plant canopy (to remove weeds and more).}

\item{areaValue}{referent value of object area in the image.}

\item{watershed}{if TRUE the "watershed" algorithm will be used to differentiate objects that are touching each other.}

\item{dissolve}{if TRUE, polygons with the same attribute value will be dissolved into multi-polygon regions. 
This option requires the rgeos package.}

\item{n.rem}{number of objects that should be removed by decreasing size (n.rem=1 is the background).}

\item{na.rm}{logical. Should missing values (including NaN) be removed?.}

\item{plot}{if TRUE the crop image and fieldShape will be plotted.}
}
\value{
A list with elements:
\itemize{
  \item \code{mosaic} (cropped by plot)
  \item \code{Dimension} (area, x.dist, y.dist)
  \item \code{numObjects} (number of objects)
  \item \code{Objects} (all objects polygon shape)
  \item \code{Polygons} (all extent polygons shape)
  \item \code{single.obj} (single object polygon shape)
  \item \code{obj.extent} (each object extent)
  \item \code{x.position} (coordinates of "x" length per object)
  \item \code{y.position} (coordinates of "y" length per object).
}
}
\description{
Calculating the object dimensions (e.g., area, "x" distance, "y" distance, number, extent, etc.) 
 in the entire mosaic or per plot using the \code{\link{fieldShape}} file.
}
\details{
fieldObject
}
