% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_fieldPlot.R
\name{fieldPlot}
\alias{fieldPlot}
\title{Plot of fieldShape file filled with trait value for each plot}
\usage{
fieldPlot(
  fieldShape,
  fieldAttribute,
  mosaic = NULL,
  color = c("white", "black"),
  min.lim = NULL,
  max.lim = NULL,
  alpha = 0.5,
  legend.position = "right",
  na.color = "gray",
  classes = 5,
  round = 3,
  horiz = FALSE
)
}
\arguments{
\item{fieldShape}{plot shape file, please use first the function getInfo().}

\item{fieldAttribute}{attribute or trait which the values will fill the plots, please use first the function getInfo().}

\item{mosaic}{object of class stack that is not necessary, but if provided will be plotted with the fieldShape file.}

\item{color}{colors to interpolate, must be a valid argument.}

\item{min.lim}{lowest limit of the color range. If is NULL the lowest value of the data will be used.}

\item{max.lim}{upper limit of the color range. If is NULL the highest value of the data will be used.}

\item{alpha}{transparency with values between 0 and 1.}

\item{legend.position}{legend position.}

\item{na.color}{color of missing values "NA".}

\item{classes}{number of classes at the legend.}

\item{round}{number of decimal digits at the legend.}

\item{horiz}{if TRUE will plot a horizontal legend.}
}
\value{
A list with two element
\itemize{
  \item The function returns a image with the \code{fieldShape} file filled with trait value for each plot.
}
}
\description{
Graphic visualization of trait values for each plot using the \code{\link{fieldShape}} file and original image.
}
\details{
fieldPlot
}
