% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_fieldRotate.R
\name{fieldRotate}
\alias{fieldRotate}
\title{Rotating the image to biuld the \code{\link{fieldShape}} file}
\usage{
fieldRotate(
  mosaic,
  theta = NULL,
  clockwise = TRUE,
  h = FALSE,
  n.core = NULL,
  extentGIS = FALSE,
  DSMmosaic = NULL,
  plot = TRUE,
  type = "l",
  lty = 2,
  lwd = 3,
  fast.plot = FALSE
)
}
\arguments{
\item{mosaic}{object of class stack.}

\item{theta}{angle of rotation, if negativo the rotation will be for the right. 
If not provided the user should select two points from left to right to determine the angle of rotation.}

\item{clockwise}{if it is TRUE, clockwise rotation.}

\item{h}{if it is TRUE, the drawn line will be at horizontal, if FALSE (90-theta).}

\item{n.core}{number of cores to use for multicore processing (Parallel).}

\item{extentGIS}{if TRUE the rotated image will have an adjusted extent 
based on the original image (It is an important step to fit the fieldShape on the original image).}

\item{DSMmosaic}{DSM should be used if the file of height is provided.}

\item{plot}{if it is TRUE the original and rotated image will be plotted.}

\item{type}{character indicating the type of plotting, please check help("lines").}

\item{lty}{line types, please check help("lines").}

\item{lwd}{line width, please check help("lines").}

\item{fast.plot}{if TRUE only the grey scale image will be plotted as reference (faster approach).}
}
\value{
A list with two element
\itemize{
  \item \code{mosaic} rotated image format stack with the base of experimental field parallel to axis X.
}
}
\description{
The image should be rotated to use the function \code{\link{fieldShape}}. 
 The base of experimental field should be parallel to axis X.
}
\details{
fieldRotate
}
