% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_imgRAST.R
\name{imgRAST}
\alias{imgRAST}
\title{Converts an EBImage object to a terra object}
\usage{
imgRAST(img, ref)
}
\arguments{
\item{img}{image/object class \code{EBImage}}

\item{ref}{SpatRaster with the crs that rast_obj should be assigned}
}
\value{
\code{terra} object
}
\description{
Converts an EBImage object to a terra object
}
\details{
imgRAST
}
