\name{Indices.txt}
\docType{data}
\title{Vegetation indices}
\description{
  List of vegetation indices avaliable at FIELDimageR.
}
\format{Data frame with NAME, FORMULA and BANDS}
\references{
  Matias FI, Caraza-Harter MV, Endelman JB. FIELDimageR: An R package to analyze orthomosaic images from agricultural field trials. The Plant Phenome J. 2020;e20005. https://doi.org/10.1002/ppj2.20005
}
\keyword{datasets}
