\name{RGB.rescale.R}
\docType{data}
\title{RGB rescale colors}
\description{
  Rescale continuous vector to have specified minimum and maximum.
}
\format{RGB.rescale(mosaic, num.band)}
\references{
  Matias et al. (2019) FIELDimageR: A R Package to Analyze Orthomosaic Images from Agricultural Field Trials. (submitted)
}
\keyword{datasets}
